Clean out-of-date function ``d_alloc_root''
``d_alloc_root'' is out of date. And it will be a compiling-time error now.
Here is the doc from kernel doc:
        d_alloc_root() is gone, along with a lot of bugs caused by code
misusing it.  Replacement: d_make_root(inode).  The difference is,
d_make_root() drops the reference to inode if dentry allocation fails.

So i change it to ``d_make_root''.